Moved all code about the camera to the fragment
