+++ app/src/main/AndroidManifest.xml
@@ -74,7 +74,6 @@
         <activity
             android:name=".MapsActivity"
             android:label="@string/title_activity_maps" />
-        <activity android:name=".TakePicActivity" />
         <activity android:name=".CameraActivity" />
         <activity android:name=".TakePicFragment" />
         <activity android:name=".UploadPicFragment"></activity>
